Refactor Page Constants
