+++ packages/bodiless-page/src/Forms/MenuFormFields.tsx
@@ -15,6 +15,12 @@
 import React from 'react';
 import { useMenuOptionUI } from '@bodiless/core';
 import { useField } from 'informed';
+import {
+  PAGE_URL_FIELD_NAME,
+  BASE_PATH_EMPTY_VALUE,
+  INPUT_FIELD_INLINE_CLASSES,
+  INPUT_FIELD_BLOCK_CLASSES,
+} from '../constants';
 import type { FieldProps } from '../types';
 import {
   fieldValueToUrl,
@@ -24,12 +30,6 @@ import {
   useBasePathField,
   validatePageUrl,
 } from '../utils';
-
-const PAGE_URL_FIELD_NAME = 'pagePath';
-const BASE_PATH_EMPTY_VALUE = '/';
-const INPUT_FIELD_DEFAULT_CLASSES = 'bl-text-gray-900 bl-bg-gray-100 bl-text-xs bl-min-w-xl-grid-1 bl-my-grid-2 bl-p-grid-1';
-const INPUT_FIELD_INLINE_CLASSES = INPUT_FIELD_DEFAULT_CLASSES.concat(' bl-inline');
-const INPUT_FIELD_BLOCK_CLASSES = INPUT_FIELD_DEFAULT_CLASSES.concat(' bl-block bl-w-full');
 
 /**
  * informed custom field that provides ability to enter new page path
